dev-split-components-to-views
* Move page facing components to views
