+++ src/Views/Favorites/_Favorites.scss
