+++ src/Views/Favorites/index.jsx
@@ -3,7 +3,7 @@ import { compose, graphql, Query } from 'react-apollo';
 import PropTypes from 'prop-types';
 import { ReactTitle } from 'react-meta-tags';
 
-import BookCard from '../BookCard';
+import BookCard from '../../Components/BookCard';
 import BookPreloader from '../BookCatalog/BookPreloader';
 
 import { toaster, errorHandler } from '../../utils';
@@ -11,7 +11,7 @@ import { removeFavorites, getFavorites } from '../../queries/books';
 import {
   SUCCESS, TOASTR_ERROR, REMOVE_FAVORITES_QUERY,
 } from '../../settings';
-import ApolloPolling from '../ApolloPolling/ApolloPolling';
+import ApolloPolling from '../../Components/ApolloPolling';
 
 class Favorites extends Component {
   state = {
@@ -129,22 +129,22 @@ class Favorites extends Component {
           startPolling, stopPolling,
         }) => (
           <Fragment>
-              <ReactTitle title="My Favorites" />
+            <ReactTitle title="My Favorites" />
 
-              {this.renderHeader()}
-              <div className="container-content">
-                {loading && <BookPreloader loadingBook={loading} />}
-                {!loading && favoriteBooks && favoriteBooks.length !== 0
-                  && this.renderFavorites(favoriteBooks)}
-                {!loading && favoriteBooks && favoriteBooks.length === 0
-                  && this.render404()}
-              </div>
+            {this.renderHeader()}
+            <div className="container-content">
+              {loading && <BookPreloader loadingBook={loading} />}
+              {!loading && favoriteBooks && favoriteBooks.length !== 0
+                && this.renderFavorites(favoriteBooks)}
+              {!loading && favoriteBooks && favoriteBooks.length === 0
+                && this.render404()}
+            </div>
 
-              <ApolloPolling
-                startPolling={startPolling}
-                stopPolling={stopPolling}
-              />
-            </Fragment>
+            <ApolloPolling
+              startPolling={startPolling}
+              stopPolling={stopPolling}
+            />
+          </Fragment>
         )}
       </Query>
     );
