+++ src/redux/constants/actionTypes.js
@@ -1,4 +1,3 @@
-/* eslint-disable */
 export const SET_LOGGEDIN_USER = 'SET_LOGGEDIN_USER';
 export const SET_BOOKS = 'SET_BOOKS';
 export const SET_LOADER = 'SET_LOADER';
