+++ src/scss/_index.scss
@@ -1,26 +1,27 @@
 @import './colors';
 @import './global';
 
+@import '../Views/AdminPanel/AdminPanel';
+@import '../Views/Base/App';
+@import '../Views/BookCatalog/BookCatalog';
+@import '../Views/BookProfile/BookProfile';
+@import '../Views/Favorites/Favorites';
+@import '../Views/MyBooks/MyBooks';
+@import '../Views/NotFound/NotFound';
+@import '../Views/PasswordReset/PasswordReset';
+@import '../Views/Users/Users';
+@import '../Views/UserProfile/ImageUpload';
+@import '../Views/UserProfile/UpdatePassword';
+@import '../Views/UserProfile/UserProfile';
+
 @import '../Components/BookCard/BookCard';
 @import '../Components/AddBook/AddBookModal';
-@import '../Components/AdminPanel/AdminPanel';
 @import '../Components/ReviewForm/ReviewForm';
-@import '../Components/Base/App';
 @import '../Components/BackToTop/BackToTop';
-@import '../Components/BookCatalog/BookCatalog';
-@import '../Components/BookProfile/BookProfile';
-@import '../Components/Favorites/Favorites';
 @import '../Components/Header/Header';
 @import '../Components/LeftSideBar/LeftSideBar';
-@import '../Components/MyBooks/MyBooks';
-@import '../Components/NotFound/NotFound';
-@import '../Components/PasswordReset/PasswordReset';
 @import '../Components/ReviewCard/ReviewCard';
 @import '../Components/ReviewCard/ReplyCard';
 @import '../Components/SideNav/SideNav';
 @import '../Components/Table/Table';
-@import '../Components/Users/Users';
-@import '../Components/UserProfile/ImageUpload';
-@import '../Components/UserProfile/UpdatePassword';
-@import '../Components/UserProfile/UserProfile';
 @import '../Components/MessageBanner/MessageBanner';
